+++ .babelrc
@@ -12,5 +12,8 @@
     ],
     "@babel/typescript",
     "@babel/react"
+  ],
+  "plugins": [
+    "@babel/plugin-proposal-class-properties"
   ]
 }
