+++ package.json
@@ -46,6 +46,7 @@
   "devDependencies": {
     "@babel/cli": "^7.0.0",
     "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.0.0",
@@ -56,14 +57,15 @@
     "babel-jest": "^23.0.1",
     "concurrently": "^3.5.1",
     "conventional-changelog-ember": "^2.0.0",
-    "docz": "^0.10.2",
-    "docz-theme-default": "^0.10.2",
     "husky": "^0.14.3",
     "jest": "^23.1.0",
+    "jest-styled-components": "^6.2.2",
     "lint-staged": "^7.2.0",
     "prettier": "^1.12.1",
     "react-dom": "^16.2.0",
+    "react-test-renderer": "^16.6.0",
     "semantic-release": "^15.5.0",
+    "static-extend": "^0.1.2",
     "styled-components": "^3.3.2",
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.13.0",
@@ -131,9 +133,6 @@
       "node_modules",
       "<rootDir>/src"
     ]
-  },
-  "directories": {
-    "example": "examples"
   },
   "keywords": [
     "react",
