--- src/__test__/Responsive.test.tsx
@@ -1,5 +1,0 @@
-describe("hi", () => {
-  it("is this working", () => {
-    console.log("hiii")
-  })
-})
